Allow ReplyToMsg to just respond to an update, not always reply
Specifically replying notifies the user, even if the chat is muted,
which is not always what we want. This allows the notifications to be
silenced by not specifically replying to a user's message, but the
function still works as a neat shortcut for sending a markdown
response to the right chat based on an update.